refactor: 사이클 제거
