+++ src/main/java/com/ktb10/munggaebe/auth/controller/OAuthController.java
@@ -1,8 +1,8 @@
 package com.ktb10.munggaebe.auth.controller;
 
-import com.ktb10.munggaebe.auth.dto.AccessTokenResponse;
-import com.ktb10.munggaebe.auth.dto.LoginResponse;
-import com.ktb10.munggaebe.auth.dto.RefreshTokenResponse;
+import com.ktb10.munggaebe.auth.service.dto.AccessTokenResponse;
+import com.ktb10.munggaebe.auth.controller.dto.LoginResponse;
+import com.ktb10.munggaebe.auth.service.dto.RefreshTokenResponse;
 import com.ktb10.munggaebe.auth.service.KakaoService;
 import com.ktb10.munggaebe.auth.service.dto.LoginDto;
 import io.swagger.v3.oas.annotations.Operation;
