+++ src/main/java/com/ktb10/munggaebe/auth/service/KakaoService.java
@@ -3,8 +3,8 @@ package com.ktb10.munggaebe.auth.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ktb10.munggaebe.auth.dto.AccessTokenResponse;
-import com.ktb10.munggaebe.auth.dto.RefreshTokenResponse;
+import com.ktb10.munggaebe.auth.service.dto.AccessTokenResponse;
+import com.ktb10.munggaebe.auth.service.dto.RefreshTokenResponse;
 import com.ktb10.munggaebe.auth.exception.OAuthResponseJsonProcessingException;
 import com.ktb10.munggaebe.auth.client.KakaoAuthClient;
 import com.ktb10.munggaebe.auth.client.KakaoMemberApiClient;
